refactor(update-weight): megatron base owns the PP adapter gather
Both megatron exporters (inkling raw, megatron-bridge) are PP-local, so
MegatronHfWeightIteratorBase implements _export_lora_named_tensors as a
PP gather around the PP-local export hook; the neutral base keeps only
validation and no longer touches training parallel state.
